# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Luna DevOps API
  version: 0.1.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/oauth/device/authorization:
    post:
      operationId: startOAuthDeviceAuthorization
      tags: [Auth]
      summary: Start an OAuth Device Authorization Grant
      x-luna-cli:
        classification: protocol-adapter
        hidden: true
        exclusionReason: Device authorization is exposed through `luna login`.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [client_id]
              properties:
                client_id: { type: string }
                scope: { type: string }
      responses:
        "200":
          description: Device and user codes for the browser verification flow.
          content:
            application/json:
              schema:
                type: object
                required: [device_code, user_code, verification_uri, verification_uri_complete, expires_in, interval]
                properties:
                  device_code: { type: string }
                  user_code: { type: string }
                  verification_uri: { type: string, format: uri }
                  verification_uri_complete: { type: string, format: uri }
                  expires_in: { type: integer, minimum: 1 }
                  interval: { type: integer, minimum: 1 }
        "400":
          description: OAuth protocol error.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OAuthProtocolError" }
...
  /api/v1/oauth/token:
    post:
      operationId: exchangeOAuthToken
      tags: [Auth]
      summary: Exchange an OAuth authorization grant for tokens
      x-luna-cli:
        classification: protocol-adapter
        hidden: true
        exclusionReason: OAuth token exchange is consumed by the CLI authentication adapter.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [grant_type]
              properties:
                grant_type:
                  type: string
                  enum:
                    - authorization_code
                    - refresh_token
                    - urn:ietf:params:oauth:grant-type:device_code
                client_id: { type: string }
                client_secret: { type: string, writeOnly: true }
                code: { type: string }
                code_verifier: { type: string }
                redirect_uri: { type: string, format: uri }
                refresh_token: { type: string, writeOnly: true }
                device_code: { type: string, writeOnly: true }
      responses:
        "200":
          description: OAuth access and refresh tokens.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OAuthTokenResponse" }
        "400":
          description: OAuth protocol error.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OAuthProtocolError" }
  /api/v1/oauth/revoke:
    post:
      operationId: revokeOAuthToken
      tags: [Auth]
      summary: Revoke an OAuth access or refresh token
      x-luna-cli:
        classification: protocol-adapter
        hidden: true
        exclusionReason: OAuth revocation is exposed through `luna logout`.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, writeOnly: true }
                token_type_hint:
                  type: string
                  enum: [access_token, refresh_token]
                client_id: { type: string }
                client_secret: { type: string, writeOnly: true }
      responses:
        "200":
          description: Token is revoked or was already invalid.
...
  /api/v1/oauth/authorize:
    get:
      operationId: getOAuthAuthorizationRequest
      tags:
        - OAuthApplications
      summary: Get OAuth Authorization Request
      x-luna-cli:
        hidden: true
        exclusionReason: Protocol endpoint consumed by a browser, webhook, probe, stream, or terminal adapter instead of a regular CLI command.
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
        - name: scope
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful business response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessObject"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: decideOAuthAuthorization
      tags:
        - OAuthApplications
      summary: Decide OAuth Authorization
      x-luna-cli:
        hidden: true
        exclusionReason: Protocol endpoint consumed by a browser, webhook, probe, stream, or terminal adapter instead of a regular CLI command.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OAuthAuthorizationDecisionInput"
      responses:
        "200":
          description: Successful business response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessObject"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: lyd_session
      description: HttpOnly browser session cookie. Sensitive interactive operations do not accept personal access tokens as a substitute.
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: Luna OAuth access token or personal access token
      description: Luna OAuth access token or personal access token. Individual operations may require OAuth and explicitly reject personal access tokens.
    AIAgentServiceToken:
      type: http
      scheme: bearer
      bearerFormat: independent Agent callback service credential
    AIDelegationToken:
      type: http
      scheme: bearer
      bearerFormat: 60-second operation-bound AI delegation
...
  responses:
...
    BadRequest:
      description: The request is invalid or cannot be processed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The authenticated principal is not allowed to perform this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        error:
          type: string
          description: Safe user-facing error message.
        detail:
          type: string
          description: Development-only diagnostic detail.
        purpose:
          type: string
          description: Step-up purpose returned with `mfa_required`.
        details:
          type: object
          additionalProperties: true
          description: Optional machine-readable error context. Scope failures include `requiredScope`.
      required: [code, error]
...
    OAuthTokenResponse:
      type: object
      properties:
        access_token: { type: string, writeOnly: true }
        token_type: { type: string, const: Bearer }
        expires_in: { type: integer, minimum: 1 }
        refresh_token: { type: string, writeOnly: true }
        scope: { type: string }
      required: [access_token, token_type, expires_in, refresh_token, scope]
    OAuthProtocolError:
      type: object
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - authorization_pending
            - slow_down
            - access_denied
            - expired_token
        error_description: { type: string }
      required: [error]
...
    BusinessObject:
      type: object
      description: Extensible business object returned by an API whose domain model is not shared across operations.
      additionalProperties: true
...
    OAuthAuthorizationDecisionInput:
      type: object
      required: [approved, clientId, redirectUri]
      properties:
        approved: { type: boolean }
        clientId: { type: string }
        redirectUri: { type: string, format: uri }
        scope: { type: string }
        state: { type: string }
        codeChallenge: { type: string }
        codeChallengeMethod: { type: string, enum: [S256] }
...